DB issues resolved

+++ src/main/java/models/FriendDto.java
@@ -1,27 +1,29 @@
 package models;
 
+import java.math.BigInteger;
+
 // import java.math.BigInteger;
 
 public class FriendDto {
     private Long id;
-    private Long userId;
-    private Long friendId;
+    private BigInteger userId;
+    private BigInteger friendId;
     public Long getId() {
         return id;
     }
     public void setId(Long id) {
         this.id = id;
     }
-    public Long getUserId() {
+    public BigInteger getUserId() {
         return userId;
     }
-    public void setUserId(Long userId) {
+    public void setUserId(BigInteger userId) {
         this.userId = userId;
     }
-    public Long getFriendId() {
+    public BigInteger getFriendId() {
         return friendId;
     }
-    public void setFriendId(Long friendId) {
+    public void setFriendId(BigInteger friendId) {
         this.friendId = friendId;
     }
